authentication cognito

+++ lambda/orders/ordersFunction.ts
@@ -1,4 +1,4 @@
-import { DynamoDB, EventBridge, SNS } from "aws-sdk"
+import { CognitoIdentityServiceProvider, DynamoDB, EventBridge, SNS } from "aws-sdk"
 import { Order, OrderRepository } from "/opt/nodejs/ordersLayer"
 import { Product, ProductRepository } from "/opt/nodejs/productsLayer"
 import * as AWSXRay from "aws-xray-sdk"
@@ -6,7 +6,7 @@ import { APIGatewayProxyEvent, APIGatewayProxyResult, Context } from "aws-lambda
 import { CarrierType, OrderProductResponse, OrderRequest, OrderResponse, PaymentType, ShippingType } from "./layers/ordersApiLayer/nodejs/orderApi"
 import { OrderEvent, OrderEventType, Envelope } from "/opt/nodejs/orderEventsLayer"
 import { v4 as uuid } from "uuid"
-
+import { AuthInfoService } from "../auth/layers/authUserInfo/nodejs/authUserInfo"
 
 AWSXRay.captureAWS(require("aws-sdk"))
 
@@ -18,9 +18,12 @@ const auditBusName = process.env.AUDIT_BUS_NAME!
 const ddbClient = new DynamoDB.DocumentClient()
 const snsClient = new SNS()
 const eventBridgeClient = new EventBridge()
+const cognitoIdentityServiceProvider = new CognitoIdentityServiceProvider()
 
 const orderRepository = new OrderRepository(ddbClient, ordersDdb)
 const productRepository = new ProductRepository(ddbClient, productsDdb)
+
+const authInfoService = new AuthInfoService(cognitoIdentityServiceProvider)
 
 export async function handler(event: APIGatewayProxyEvent, context: Context): Promise<APIGatewayProxyResult> {
     const method = event.httpMethod
@@ -29,49 +32,79 @@ export async function handler(event: APIGatewayProxyEvent, context: Context): Pr
 
     console.log(`API Gateway RequestId: ${apiRequestId} - LambdaRequestId :${lambdaRequestId}`)
 
+    const isAdmin = authInfoService.isAdminUser(event.requestContext.authorizer)
+    const authenticatedUser = await authInfoService.getUserInfo(event.requestContext.authorizer)
+
     if (method === 'GET') {
         if (event.queryStringParameters) {
             const email = event.queryStringParameters!.email
             const orderId = event.queryStringParameters!.orderId
-            if (email) {
-                if (orderId) {
-                    // Get one order from an user
-                    try {
-                        const order = await orderRepository.getOrder(email, orderId)
+
+            if (isAdmin || email === authenticatedUser) {
+                if (email) {
+                    if (orderId) {
+                        // Get one order from an user
+                        try {
+                            const order = await orderRepository.getOrder(email, orderId)
+                            return {
+                                statusCode: 200,
+                                body: JSON.stringify(convertToOrderResponse(order))
+                            }
+                        } catch (error) {
+                            console.log((<Error>error).message)
+                            return {
+                                statusCode: 404,
+                                body: (<Error>error).message
+                            }
+                        }
+
+
+                    } else {
+                        // Get all orders from an user
+                        const orders = await orderRepository.getOrdersByEmail(email)
                         return {
                             statusCode: 200,
-                            body: JSON.stringify(convertToOrderResponse(order))
+                            body: JSON.stringify(orders.map(convertToOrderResponse))
                         }
-                    } catch (error) {
-                        console.log((<Error>error).message)
-                        return {
-                            statusCode: 404,
-                            body: (<Error>error).message
-                        }
-                    }
-
-
-                } else {
-                    // Get all orders from an user
-                    const orders = await orderRepository.getOrdersByEmail(email)
-                    return {
-                        statusCode: 200,
-                        body: JSON.stringify(orders.map(convertToOrderResponse))
                     }
                 }
+            } else {
+                return {
+                    statusCode: 403,
+                    body: `You don't have permission to access this operation`
+                }
             }
+
         } else {
             // Get all orders
-            const orders = await orderRepository.getAllOrders()
-            return {
-                statusCode: 200,
-                body: JSON.stringify(orders.map(convertToOrderResponse))
+            if (isAdmin) {
+                const orders = await orderRepository.getAllOrders()
+                return {
+                    statusCode: 200,
+                    body: JSON.stringify(orders.map(convertToOrderResponse))
+                }
+            } else {
+                return {
+                    statusCode: 403,
+                    body: `You don't have permission to access this operation`
+                }
             }
         }
+
 
     } else if (method === 'POST') {
         console.log('POST /orders')
         const orderRequest = JSON.parse(event.body!) as OrderRequest
+
+        if (!isAdmin) {
+            orderRequest.email = authenticatedUser
+        } else if (orderRequest.email === null) {
+            return {
+                statusCode: 400,
+                body: 'Missing the order owner email'
+            }
+        }
+
         const products = await productRepository.getProductsByIds(orderRequest.productIds)
         if (products.length === orderRequest.productIds.length) {
             const order = buildOrder(orderRequest, products)
@@ -91,6 +124,8 @@ export async function handler(event: APIGatewayProxyEvent, context: Context): Pr
                 body: JSON.stringify(convertToOrderResponse(order))
             }
         } else {
+            console.error('Some product was not found')
+
             const result = await eventBridgeClient.putEvents({
                 Entries: [
                     {
@@ -119,27 +154,33 @@ export async function handler(event: APIGatewayProxyEvent, context: Context): Pr
         const email = event.queryStringParameters!.email!
         const orderId = event.queryStringParameters!.orderId!
 
-        try {
-            const orderDelete = await orderRepository.deleteOrder(email, orderId)
+        if (isAdmin || email === authenticatedUser) {
+            try {
+                const orderDelete = await orderRepository.deleteOrder(email, orderId)
 
-            const eventResult = await sendOrderEvent(orderDelete, OrderEventType.DELETED, lambdaRequestId)
-            console.log(
-                `Order deleted event sent - OrderId: ${orderDelete.sk}
-                - MessageId: ${eventResult.MessageId}`
-            )
+                const eventResult = await sendOrderEvent(orderDelete, OrderEventType.DELETED, lambdaRequestId)
+                console.log(
+                    `Order deleted event sent - OrderId: ${orderDelete.sk}
+                    - MessageId: ${eventResult.MessageId}`
+                )
 
-            return {
-                statusCode: 200,
-                body: JSON.stringify(convertToOrderResponse(orderDelete))
+                return {
+                    statusCode: 200,
+                    body: JSON.stringify(convertToOrderResponse(orderDelete))
+                }
+            } catch (error) {
+                console.log((<Error>error).message)
+                return {
+                    statusCode: 404,
+                    body: (<Error>error).message
+                }
             }
-        } catch (error) {
-            console.log((<Error>error).message)
+        } else {
             return {
-                statusCode: 404,
-                body: (<Error>error).message
+                statusCode: 403,
+                body: `You don't have permission to access this operation`
             }
         }
-
 
     }
 
@@ -194,7 +235,7 @@ function convertToOrderResponse(order: Order): OrderResponse {
         email: order.pk,
         id: order.sk!,
         createdAt: order.createdAt!,
-        products: orderProducts. length ? orderProducts : undefined, 
+        products: orderProducts.length ? orderProducts : undefined,
         billing: {
             payment: order.billing.payment as PaymentType,
             totalPrice: order.billing.totalPrice
